Move table_factory to top-level conftest
